Dash moves through walls
Currently, the dash moves through walls. This is because of the MovePosition/Lerp method of moving the player when they dash. I will have to come up with an alternative in the future, but for now, this will remain. It will require some thought, or use of a .AddForce method.