Player takes damage, game over on 0 HP
The player can now take damage, and everything will stop when the character is at 0HP. There are no current ingame indicators so please check the inspector. The player object's hierarchy was reshuffled a bit, so I will need to double check that nothing will cause errors in the future.