Rotated input vector for movement animation, moved animations to AvatarAspect
The animations for moving in different directions were not transitioning properly. This was leading to the model running right, but looking like it was running left after a certain point. This has now been fixed. It also made more sense to have the AvatarAspect have the responsibility of animating the model.