Physics based dash
Switched the dash to be physics based, but neglected to make a separate branch. Undid some work due to conflicts. This is about halfway to where I was at. The Barrage should hang just long enough to fire, then start to drop. There should be a brief period after where the player is in active state, but no inputs are accepted.