Aiming Arc and Ring, Blast/barrage movement in air, no super sliding while blasting on ground
Worked way too hard to get the trajectory a nice curve, a simple triangle will do for now. When grounded will project a path of projectile, when in air will project a straight line. Aiming ring moves via the right stick when the right trigger is held down. It spins slightly. Made it so that only upwards velocity is canceled when firing a barrage/blast. Prevented the player from being able to aim blast and move briefly at full speed through sliding.