Blast emitter for all avatars
Added a blast and barrage emitter object so that when projectiles are spawned, they start in a spot corresponding to where the avatar's hand is. There was also the issue of projectiles hitting the Avatar and exploding, or the aiming ring behaving oddly due to it being a child. This addresses both.