State overwriting cleared up
There was an issue when spamming air dash and blast, the player would get stuck in the air. This is mostly cleared up. I've been able to replicate it a few times after the fix, but I'm not sure of the cause. Also, if a player is invincible near something that can cause an explosion, they can spam their blast at a high rate. This should at least be capped in a future update.